Added binding AirPlane -> AirPlane in Hangar <- Hangar Finding Airplane by code Park Airplane into Hangar

+++ src/main/java/com/cleverlance/academy/airlines/AirlinesApplication.java
@@ -23,7 +23,7 @@ public class AirlinesApplication {
 				.numOfSeatsInEC(123)
 				.build());
 		hSer.createHangar(Hangar.builder()
-				.id(1)
+				.id((long) 1)
 				.name("Pavlovo")
 				.build());
